refactor: Fix PauseGame script and Game Scene UI
