+++ Assets/Scenes/MainMenu.unity
@@ -862,7 +862,7 @@ GameObject:
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
   m_StaticEditorFlags: 0
-  m_IsActive: 0
+  m_IsActive: 1
 --- !u!224 &404954239
 RectTransform:
   m_ObjectHideFlags: 0
