--- Assets/Scripts/PauseGame/PauseGame.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 0e4d0dd783e76ef46a06a5234f12f90d
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
