+++ ProjectSettings/EditorBuildSettings.asset
@@ -11,18 +11,6 @@ EditorBuildSettings:
   - enabled: 1
     path: Assets/Scenes/MainMenu.unity
     guid: d3f659f39e90a364199ee6294a0c4ca4
-  - enabled: 0
-    path: Assets/Scenes/Game Menu.unity
-    guid: 8c9cfa26abfee488c85f1582747f6a02
-  - enabled: 0
-    path: Assets/Scenes/Load Game.unity
-    guid: 5630d44231247b243a8e11a08aa1c574
-  - enabled: 0
-    path: Assets/Scenes/Options.unity
-    guid: 2749686ea36394f479dc4a38cf706de4
-  - enabled: 0
-    path: Assets/Scenes/Sound Settings.unity
-    guid: 894a846c97dbb3943b2d95e3374011f0
   - enabled: 1
     path: Assets/Scenes/GameScene.unity
     guid: 462e8a939e54347438259958655e5348
